fixes #146
Non-generic IInlineContentItemsResolver was removed as it was just an unusable interface.
An independent ITypelessInlineContentItemsResolver interface was introduced to wrap
IInlineContentItemsResolver<TContentItem> resolvers and provide them in a
non-generic way to the InlineContentItemsProcessor that later puts inline
content items through the resolver in order to obtain their string representation.

New extension methods for IServiceCollection were introduced.